fix up expiration parameters

+++ Queue.ExpireNodes.go
@@ -3,12 +3,18 @@ package expiringQueue
 import "time"
 
 // expireNodes - run in the background to expire nodes by age.
-func (q *Queue[T]) expireNodes(expiration int64) {
-	q.expiration = expiration
+func (q *Queue[T]) expireNodes() {
+	if q.expiration == 0 {
+		return
+	}
 	for {
 		var prev *Node[T]
 		time.Sleep(time.Duration(q.expiration) * time.Nanosecond)
 		q.lock.Lock()
+		if q.pauseExpirationTask {
+			q.lock.Unlock()
+			return
+		}
 		now := time.Now().UnixNano()
 		for node := q.head; node != nil; node = node.next {
 			if age := node.timestamp - now; age > q.expiration && q.deleteNode(node, prev) {
